Refactor project files to be much more cleaner
