+++ requirements.txt
@@ -1,4 +1,3 @@
+openai==1.2.0
 Pillow==10.1.0
 requests==2.31.0
-openai==1.2.0
-
